requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.5.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
fastapi==0.100.1
filelock==3.15.4
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth-httplib2==0.2.0
google-auth==2.34.0
googleapis-common-protos==1.63.2
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
identify==2.6.0
idna==3.7
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.7.1
proto-plus==1.24.0
protobuf==5.27.3
pyasn1-modules==0.4.0
pyasn1==0.6.0
pydantic-core==2.20.1
pydantic==2.8.2
pyparsing==3.1.4
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.27.0
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn[standard]==0.23.2
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.22.0
websockets==12.0
